Move merging functionality to helper in Entity
